# services/operator-interface/src/OperatorInterface.Api/Adapters/Http/openapi.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Operator Interface
  description: API для работы управления сессией работы оператора
paths:
  /api/v1/operator/sessions/authorize:
    post:
      summary: Создать новую сессию оператора
      description: Позволяет создать новую сессию для работы оператора
      operationId: Authorize
      requestBody:
        description: Курьер
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'      
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SessionId'          
        default:
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
                
  /api/v1/operator/sessions/{sessionId}:
    get:
      summary: Получить сессию 
      description: Позволяет получить корзину сессию оператора
      operationId: GetSession
      parameters:
        - name: sessionId
          in: path
          description: Идентификатор сессии
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
                
        default:
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Credentials:
      type: object
      required:
        - login
        - password
        - workplace
      properties:
        login:
          type: string
          description: Логин
        password:
          type: string
          description: Пароль
        workplace:
          type: string
          description: Код окна
    SessionId:
      type: object
      required:
        - value
      properties:
        value:
          type: string
          format: uuid
          description: Идентификатор сессии
          
    Session:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Идентификатор сессии
        status:
          type: string
          description: Статус
          enum:
            - created
            - confirmed
  
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: Код ошибки
        message:
          type: string
          description: Текст ошибки